cloudformationrepo/requirements.txt
aniso8601==9.0.1
astroid==2.5.6
Babel==2.9.1
boto3==1.17.66
botocore==1.20.66
click==7.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-RESTful==0.3.8
Flask-Table==0.5.0
gitdb==4.0.7
GitPython==3.1.14
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.8.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
s3transfer==0.4.2
six==1.15.0
smmap==4.0.0
toml==0.10.2
urllib3==1.26.4
uwsgi==2.0.19.1
Werkzeug==1.0.1
wrapt==1.12.1
